Consolidate smoke tests and tighten test hygiene
